+++ packages/python/m/ci/celt/post_processor.py
@@ -3,7 +3,7 @@ from typing import Dict
 from ...core import Good, Res, issue
 from .core.process import PostProcessor
 from .core.types import Configuration, Transform
-from .post_processors import eslint, pycodestyle, pylint, typescript
+from .post_processors import eslint, pycodestyle, pylint, ruff, typescript
 
 
 def get_post_processor(
@@ -25,6 +25,7 @@ def get_post_processor(
         'flake8': pycodestyle.read_payload,
         'pylint': pylint.read_payload,
         'typescript': typescript.read_payload,
+        'ruff': ruff.read_payload,
     }
     if name not in mapping:
         return issue(f'{name} is not a supported post processor')