+++ packages/python/tests/run.sh
@@ -13,7 +13,7 @@ else
 # To run specific tests:
 # python -m unittest discover -s packages/python -v -k tests.cli.commands.test_json.CliJsonTest
 # python -m pytest -vv -k test_m_npm
-  pytest -p no:logging packages/python -vv -k test_pydantic
+  pytest -p no:logging packages/python -vv -k test_m_ci_celt[tcase3]
 fi
 
 mkdir -p m/.m